% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL.R
\docType{data}
\name{AL}
\alias{AL}
\title{Simulated 3-layer network data}
\format{
A list of length 3
}
\usage{
AL
}
\description{
A data set containing 3 layers of undirected, symmetric adjacency matrices
simulated from an SBM with 3 true clusters
}
\keyword{datasets}
