% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsbm.R
\docType{package}
\name{mlsbm}
\alias{mlsbm}
\title{mypackage: A package for fitting single and multilevel SBMs.}
\description{
This package fits Bayesian stochastic block models (SBMs)
}
\section{mlsbm functions}{

The mlsbm functions ...
}

