% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlsbm.R
\name{fit_mlsbm}
\alias{fit_mlsbm}
\title{R/Rcpp function for fitting multilevel stochastic block model}
\usage{
fit_mlsbm(
  A,
  K,
  z_init = NULL,
  a0 = 2,
  b10 = 1,
  b20 = 1,
  n_iter = 1000,
  burn = 100,
  verbose = FALSE,
  r = 1.2
)
}
\arguments{
\item{A}{An adjacency list of length L, the number of levels. Each level contains an n x n symmetric adjacency matrix.}

\item{K}{The number of clusters specified a priori.}

\item{z_init}{Initialized cluster indicators. If NULL, will initialize automatically with Louvain algorithm.}

\item{a0}{Dirichlet prior parameter for cluster sizes for clusters 1,...,K.}

\item{b10}{Beta distribution prior paramter for community connectivity.}

\item{b20}{Beta distribution prior parameter for community connectivity.}

\item{n_iter}{The number of total MCMC iterations to run.}

\item{burn}{The number of burn-in MCMC iterations to discard. The number of saved iterations will be n_iter - burn.}

\item{verbose}{Whether to print a progress bar to track MCMC progress. Defaults to true.}

\item{r}{Resolution parameter for Louvain initialization. Sould be >= 0 and higher values give a larger number of smaller clusters.}
}
\value{
A list of MCMC samples, including the MAP estimate of cluster indicators (z)
}
\description{
This function allows you to fit multilevel stochastic block models.
}
\examples{
data(AL)
# increase n_iter in practice
fit <- fit_mlsbm(AL,3,n_iter = 100)
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{networks}
